#19-feat/확장 input을 공동 모듈로 추출
